requirements.txt
PySoundFile
scipy~=1.6.2